Add files via upload
Wall Blaster 2 Enemy

The Wall Blaster II enemy was requested a while ago but I didn't have any time to work on it until now. The project doesn't make use of the tutorial series dependencies so it should be a lot easier to integrate into an existing project. There are three colored sprites and three cannon positions so in total nine variations. I set the shoot delay to 1.5 seconds (approximate timing from MM6 as I think it varied). WB2 doesn't shoot when the Player is behind the cannon but when in front up to an attack distance you can set. You can change the orientation between left and right facing. Also included a bunch of public methods in WB2's script so you can change him up during runtime.

https://youtu.be/I7E6b6P3o6s

Made with Unity 2019.3.10f1